pyproject.toml
[project]
name = "splitbill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "django==5.2.4",
    "django-extensions==4.1",
    "django-rest-passwordreset==1.5.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular==0.28.0",
    "pip-system-certs==5.2",
    "pre-commit==4.3.0",
    "psycopg[binary]==3.2.9",
    "pylint-django==2.6.1",
    "python-decouple==3.8",
    "requests==2.32.4",
]
[tool.ruff.lint]
select = ["F"]
ignore = ["F841"]
[tool.pyright]
typeCheckingMode = "off"